refactor: split models.py into modular models package
