+++ models/common.py
@@ -1,0 +1,62 @@
+from enum import Enum
+
+
+class CustomerType(str, Enum):
+    individual = "individual"
+    company = "company"
+
+
+class LeadStatus(str, Enum):
+    new = "new"
+    contacted = "contacted"
+    qualified = "qualified"
+    lost = "lost"
+    spam = "spam"
+
+
+class LeadIntakeSource(str, Enum):
+    phone = "phone"
+    site = "site"
+    bot = "bot"
+    email = "email"
+    manager = "manager"
+    other = "other"
+
+
+class LeadSegmentHint(str, Enum):
+    unknown = "unknown"
+    b2c = "b2c"
+    b2b = "b2b"
+
+
+class LeadLossReason(str, Enum):
+    no_product = "no_product"
+    no_budget = "no_budget"
+    no_response = "no_response"
+    duplicate = "duplicate"
+    spam = "spam"
+    other = "other"
+
+
+class OrderStatus(str, Enum):
+    NEW_LEAD = "new_lead"
+    ASSESSMENT = "assessment"
+    PROPOSAL = "proposal"
+    NEGOTIATION = "negotiation"
+    DEFERRED = "deferred"
+    WON_DEPOSIT = "won_deposit"
+    INSTALLATION = "installation"
+    COMPLETED = "completed"
+    CANCELED = "canceled"
+
+
+class LeadSource(str, Enum):
+    """Источник лида - откуда пришёл клиент"""
+
+    SITE = "site"
+    BOT = "bot"
+    PHONE = "phone"
+    EMAIL = "email"
+    MANAGER = "manager"
+    REFERRAL = "referral"
+    OTHER = "other"
